+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 売上レポートを予約以外の注文アイテムに対応
+- 売上レポートのusernameを注文のbrokerに対応
 
 ### Deprecated